# Implement comprehensive monitoring stack with Grafana, Loki, and Prometheus