--- a/.cursorindexingignore
+++ b/.cursorindexingignore
@@ -1,0 +1,2 @@
+docs/archive
+docs/plans/completed