--- a/grafana/grafana.ini
+++ b/grafana/grafana.ini
@@ -1,0 +1,33 @@
+[server]
+http_port = 3000
+domain = 10.1.6.12
+
+[security]
+admin_user = admin
+admin_password = ${GF_SECURITY_ADMIN_PASSWORD}
+disable_gravatar = true
+cookie_secure = true
+cookie_samesite = strict
+
+[users]
+allow_sign_up = false
+allow_org_create = false
+auto_assign_org = true
+auto_assign_org_role = Viewer
+
+[auth.anonymous]
+enabled = false
+
+[log]
+mode = console
+level = info
+
+[paths]
+data = /var/lib/grafana
+logs = /var/log/grafana
+plugins = /var/lib/grafana/plugins
+provisioning = /etc/grafana/provisioning
+
+[provisioning]
+datasources = /etc/grafana/provisioning/datasources
+dashboards = /etc/grafana/provisioning/dashboards