--- a/grafana/grafana.ini
+++ b/grafana/grafana.ini
@@ -1,6 +1,9 @@
 [server]
 http_port = 3000
-domain = 10.1.6.12
+domain = webhost.hccc.edu
+root_url = https://webhost.hccc.edu/
+enable_gzip = true
+cors_allow_credentials = true
 
 [security]
 admin_user = admin
@@ -17,6 +20,12 @@ auto_assign_org_role = Viewer
 
 [auth.anonymous]
 enabled = false
+
+[cors]
+allow_credentials = true
+allowed_origins = https://webhost.hccc.edu,https://10.1.6.12,https://grafana.10.1.6.12
+allowed_headers = Accept,Authorization,Cache-Control,Content-Type,DNT,If-Modified-Since,Keep-Alive,Origin,User-Agent,X-Requested-With
+allowed_methods = GET,POST,OPTIONS,PUT,DELETE
 
 [log]
 mode = console