## Summary
This PR implements a complete monitoring and observability stack for the QR system, fulfilling Task M.3 of the Observatory-First refactoring plan. The implementation includes metrics collection, log aggregation, and comprehensive dashboards for system monitoring.

## Changes Made

### 🔧 Infrastructure & Configuration
- **Docker Compose**: Added Grafana, Loki, Promtail, and Prometheus services
- **Traefik Routing**: Configured secure access to Grafana dashboard via internal IP
- **Prometheus**: Set up metrics collection with proper service discovery
- **Loki**: Implemented centralized log aggregation system
- **Promtail**: Configured log shipping from application containers

### 📊 Monitoring Dashboards
Created three comprehensive Grafana dashboards:

1. **QR System Health Overview**
   - System resource monitoring (CPU, memory, disk)
   - Container health status
   - Network metrics
   - Alert thresholds for critical metrics

2. **QR System Refactoring Progress**
   - Code quality metrics tracking
   - Test coverage monitoring
   - Deployment frequency
   - Error rate trends

3. **QR Analytics Deep Dive**
   - QR code usage analytics
   - Geographic distribution of scans
   - Popular QR codes tracking
   - User engagement metrics

### 🔐 Security & Access Control
- Grafana accessible only via internal IP (10.1.6.12)
- Basic authentication for dashboard access
- IP whitelisting for monitoring services
- Secure TLS configuration

### 📁 File Changes
- `docker-compose.yml`: Added monitoring services with proper networking
- `dynamic_conf.yml`: Configured Traefik routing for Grafana
- `prometheus.yml`: Metrics collection configuration
- `grafana/`: Dashboard definitions and datasource configurations
- `loki/`: Log aggregation service configuration
- `promtail/`: Log shipping configuration
- `.gitignore`: Added monitoring service directories
- `memory-bank/`: Project documentation and context

## Testing
- ✅ All services start successfully via Docker Compose
- ✅ Grafana dashboards load and display data
- ✅ Prometheus metrics collection verified
- ✅ Loki log aggregation functional
- ✅ Traefik routing working correctly
- ✅ Authentication and access controls tested

## Deployment Notes
- Services are configured to restart automatically
- Persistent volumes ensure data retention
- Monitoring stack is isolated in dedicated Docker network
- Resource limits set to prevent system overload

## Related Issues
- Closes: Task M.3 - Observatory-First Monitoring Implementation
- Related: Observatory-First refactoring plan

## Breaking Changes
None. This is purely additive functionality.

## Post-Deployment Steps
1. Access Grafana at `https://10.1.6.12/grafana`
2. Login with configured credentials
3. Verify all dashboards are loading data
4. Set up alerting rules as needed
5. Configure retention policies for logs and metrics